requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.6.0
gitpython==3.1.43
httpx==0.27.0
pydantic==2.10.6
pydantic-settings==2.6.1
python-dotenv==1.0.1
